LLM_APP_DEMO/requirements.txt
beautifulsoup4
fastapi[standard]
uvicorn[standard]
langchain
langchain-community
langchain-google-genai
langchain-mongodb
langchain-text-splitters
langchainhub
python-dotenv
pymongo
requests
